epoc/services/skn: move store gfx into chunk mter
