+++ src/emu/epoc/src/services/akn/skin/server.cpp
@@ -206,8 +206,6 @@ namespace eka2l1 {
 
         fbss = reinterpret_cast<fbs_server*>(&(*kern->get_by_name<service::server>("!Fontbitmapserver")));
 
-        bitmap_store_ = std::make_unique<epoc::akn_skin_bitmap_store>();
-
         // Create skin chunk
         skin_chunk_ = kern->create_and_add<kernel::chunk>(kernel::owner_type::kernel,
             sys->get_memory_system(), nullptr, "AknsSrvSharedMemoryChunk",
@@ -241,62 +239,6 @@ namespace eka2l1 {
         fbsbitmap *bmp = fbss->get<fbsbitmap>(bmp_handle);
         fbsbitmap *msk = msk_handle ? fbss->get<fbsbitmap>(msk_handle) : nullptr;
 
-        bitmap_store_->store_bitmap(bmp);
-        if (msk) {
-            bitmap_store_->store_bitmap(msk);
-        }
-    
-        const epoc::skn_scalable_item_def def {
-            item_id,
-            bmp->id,
-            msk ? msk->id : 0,
-            layout_info.layout_type,
-            false,
-            static_cast<std::uint64_t>(time(0)),
-            layout_info.layout_size
-        };
-        const std::uint8_t *new_data = reinterpret_cast<const std::uint8_t*>(&def);
-
-        // store into shared chunk
-        epoc::skn_scalable_item_def *table = reinterpret_cast<epoc::skn_scalable_item_def*>(
-            chunk_maintainer_->get_area_base(epoc::akn_skin_chunk_area_base_offset::gfx_area_base));
-        const std::size_t gfx_area_size = chunk_maintainer_->get_area_size(epoc::akn_skin_chunk_area_base_offset::gfx_area_base);
-        const std::size_t gfx_current_size = chunk_maintainer_->get_area_current_size(epoc::akn_skin_chunk_area_base_offset::gfx_area_base);
-        const std::size_t def_count = gfx_current_size / sizeof(def);
-        // update if exist
-        for (std::size_t i = 0; i < def_count; i++) {
-            if (table[i].item_id == item_id
-            && table[i].layout_type == layout_info.layout_type
-            && table[i].layout_size == layout_info.layout_size) {
-                bitmap_store_->remove_stored_bitmap(table[i].bitmap_handle);
-                bitmap_store_->remove_stored_bitmap(table[i].mask_handle);
-                std::uint8_t *dest = reinterpret_cast<std::uint8_t*>(table + i);
-                std::copy(new_data, new_data + sizeof(def), dest);
-                return;
-            }
-        }
-        if (def_count >= gfx_area_size / sizeof(def)) {
-            // replace the oldest one if the chunk is full
-            std::uint64_t oldest_timestamp = table[0].item_timestamp;
-            std::size_t oldest_index = 0;
-            for (std::size_t i = 0; i < def_count; i++) {
-                if (table[i].item_timestamp < oldest_timestamp) {
-                    oldest_timestamp = table[i].item_timestamp;
-                    oldest_index = i;
-                }
-            }
-            bitmap_store_->remove_stored_bitmap(table[oldest_index].bitmap_handle);
-            bitmap_store_->remove_stored_bitmap(table[oldest_index].mask_handle);
-            std::uint8_t *dest = reinterpret_cast<std::uint8_t*>(table + oldest_index);
-            std::copy(new_data, new_data + sizeof(def), dest);
-        } else {
-            // add a new one
-            std::uint8_t *dest = reinterpret_cast<std::uint8_t*>(table + def_count);
-            chunk_maintainer_->set_area_current_size(
-                epoc::akn_skin_chunk_area_base_offset::gfx_area_base,
-                gfx_current_size + sizeof(def)
-            );
-            std::copy(new_data, new_data + sizeof(def), dest);
-        }
+        chunk_maintainer_->store_scalable_gfx(item_id, layout_info, bmp, msk);
     }
 }
